moves usage test to the functions tests
